requirements.txt
groq
openai
pdfplumber
pydantic
python-dotenv
rich
click
